refactor: move Sentry files to lib/sentry/ and document lib/ vs utils/ convention
Sentry is a third-party integration with initialization lifecycle and
consent management — it belongs in lib/ (configured third-party wrappers),
not utils/ (pure utility functions).

Also updates CONVENTIONS.md with a clear decision table for lib/ vs utils/
placement so future contributors know the distinction.
